split types/joints and rename imports
